pyproject.toml
[project]
name = "musnad-ocr-tool"
version = "1.0.0"
description = "Comprehensive web application for preparing OCR training data specifically for Ancient South Arabian Musnad script"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "werkzeug>=3.1.3",
]
